Make Device fields public
A private field has also been added so that a Device can only be
created through the constructor (via TryFrom).

This enables the name() method to return a &str rather than a
Result, since it can be guaranteed that one of the fields has
Some(value) and therefore unwrap() won't panic.

The Debug trait is also implemented for Device so that the private
field is not shown.